Implementation of HW-Project2
For use with PIC18F47K42. Creates a counter (0 -> F) on a 7 Segment Display, it then increments or decrements depending on the button pressed or held. If both buttons are pressed it is reset to 0. If neither button is pressed, no changes occur.

Eventual implementation with a Shaking Position Switch is planned, such that the device can perform as a Pedometer.